Fix CSS reference error, add some button handling and code to display user input
Changed CSS to use ID, as this will be unique, so class was less suitable. Button handling added to start taking some user input. Code displays user input on the calculator screen but still needs a working limiter to stop overflow. Text direction RTL changed to text-align: right; as numbers/characters were not behaving as expected as RTL is designed for languages that read RTL.